+++ qualtran/bloqs/basic_gates/__init__.py
@@ -23,6 +23,7 @@ requirements.
 
 from .cnot import CNOT
 from .hadamard import Hadamard
+from .on_each import OnEach
 from .rotation import Rx, Ry, Rz, XPowGate, YPowGate, ZPowGate
 from .swap import CSwap, TwoBitCSwap, TwoBitSwap
 from .t_gate import TGate
