+++ qualtran/bloqs/basic_gates/on_each.py
