+++ qualtran/bloqs/rotations/phase_gradient.py
@@ -24,8 +24,8 @@ from numpy.typing import NDArray
 
 from qualtran import GateWithRegisters, QBit, QFxp, Register, Side, Signature
 from qualtran.bloqs.basic_gates import Hadamard, Toffoli
+from qualtran.bloqs.basic_gates.on_each import OnEach
 from qualtran.bloqs.basic_gates.rotation import CZPowGate, ZPowGate
-from qualtran.bloqs.on_each import OnEach
 
 if TYPE_CHECKING:
     from qualtran.resource_counting.bloq_counts import BloqCountT
